# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amane API",
    "version": "0.7.2"
  },
  "paths": {
...
    "/api/schedules": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "List Schedules",
        "operationId": "list_schedules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Create Schedule",
        "operationId": "create_schedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/schedules/{schedule_id}": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "Get Schedule",
        "operationId": "get_schedule",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "schedules"
        ],
        "summary": "Update Schedule",
        "description": "仅 name/cron/enabled. 修改任务内容须删除后重建.",
        "operationId": "update_schedule",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "schedules"
        ],
        "summary": "Delete Schedule",
        "operationId": "delete_schedule",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules/{schedule_id}/trigger": {
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Trigger Schedule",
        "description": "将 next_run 设置为当前时间, 由 cron 在下一个 tick 执行.",
        "operationId": "trigger_schedule",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CleanupSubmission": {
        "properties": {
          "remove_missing_files": {
            "type": "boolean",
            "title": "Remove Missing Files",
            "description": "删除磁盘上不存在的 MediaFile 记录 (不触碰 Metadata)",
            "default": true
          },
          "remove_unreferenced_resources": {
            "type": "boolean",
            "title": "Remove Unreferenced Resources",
            "description": "删除不被任何 Metadata URL 字段引用的 Resource (含派生裁剪)",
            "default": true
          },
          "type": {
            "type": "string",
            "const": "cleanup",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "CleanupSubmission"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "R18ImportSubmission": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "type": {
            "type": "string",
            "const": "r18_import",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "R18ImportSubmission"
      },
...
      "RescrapeSubmission": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "单次最多补刮的元数据数 (避免长占 worker 队列)",
            "default": 100
          },
          "min_age_days": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Age Days",
            "description": "仅补刮 updated_at 距今超过该天数的条目; None 不设门槛"
          },
          "type": {
            "type": "string",
            "const": "rescrape",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "RescrapeSubmission"
      },
      "RoutineType": {
        "type": "string",
        "enum": [
          "cleanup",
          "upscale",
          "r18_import",
          "rescrape"
        ],
        "title": "RoutineType"
      },
...
      "ScheduleCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cron": {
            "type": "string",
            "title": "Cron"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "submission": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CleanupSubmission"
              },
              {
                "$ref": "#/components/schemas/UpscaleSubmission"
              },
              {
                "$ref": "#/components/schemas/R18ImportSubmission"
              },
              {
                "$ref": "#/components/schemas/RescrapeSubmission"
              }
            ],
            "title": "Submission",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "cleanup": "#/components/schemas/CleanupSubmission",
                "r18_import": "#/components/schemas/R18ImportSubmission",
                "rescrape": "#/components/schemas/RescrapeSubmission",
                "upscale": "#/components/schemas/UpscaleSubmission"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "cron",
          "submission"
        ],
        "title": "ScheduleCreateRequest"
      },
      "ScheduleListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScheduleResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ScheduleListResponse"
      },
      "ScheduleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cron": {
            "type": "string",
            "title": "Cron"
          },
          "task_type": {
            "$ref": "#/components/schemas/RoutineType"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "last_run": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run"
          },
          "next_run": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cron",
          "task_type",
          "enabled"
        ],
        "title": "ScheduleResponse"
      },
      "ScheduleUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cron": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ScheduleUpdateRequest"
      },
...
      "UpscaleSubmission": {
        "properties": {
          "max_dim_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Dim Threshold"
          },
          "max_bytes_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Bytes Threshold"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 200
          },
          "type": {
            "type": "string",
            "const": "upscale",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "UpscaleSubmission"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}